Fix clippy warnings
- Use div_ceil() instead of manual implementation
- Remove unnecessary cast
- Add Default implementations for GearHardware and GearScore
- Use Range::contains() for bounds checks
- Add safety doc to DllMain
- Use c-string literals instead of manual null-terminated strings

+++ src/gear_graphics.rs
@@ -72,7 +72,7 @@ unsafe extern "thiscall" fn gear_graphics_get_desktop_resolution(
         *height = (rect.bottom - rect.top) as u32;
     }
 
-    (rect.bottom - rect.top) as i32
+    rect.bottom - rect.top
 }
 
 impl GearGraphics {
